libzeth: move circuit-related includes into circuits directory
