+++ prover_server/prover_server.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: LGPL-3.0+
 
-#include "libzeth/circuit_types.hpp"
+#include "libzeth/circuits/circuit_types.hpp"
 #include "libzeth/serialization/api/api_io.hpp"
 #include "libzeth/serialization/file_io.hpp"
 #include "libzeth/snarks/default/default_api.hpp"
